ref : 도메인에 맞게 위치 변경
